% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craggy2019.R
\docType{package}
\name{craggy2019}
\alias{craggy2019}
\alias{craggy2019-package}
\title{craggy2019: Housing and eviction data for King County, Washington, United States.}
\description{
This package provides data, as well as some utilities for accessing that
data. See the \code{vignettes} for sample code for exploring and analyzing
the data.
}
