% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_king_pit.R
\name{get_king_pit}
\alias{get_king_pit}
\title{Counts PIT per year for King County.}
\usage{
get_king_pit()
}
\value{
data.frame with one row per year.
}
\description{
Counts PIT per year for King County.
}
